Refactored arrows component, adding tests now
