--- src/components/atoms/calendar-arrow-right/index.js
@@ -1,23 +1,0 @@
-import styles from "./calendar-arrow-right.module.css";
-import { ChevronDown } from "../chevron-down";
-
-const RightCalendarArrow = ({ containerRef, cardWidth, clickedYear }) => {
-  const scrollWidth = cardWidth?.offsetWidth;
-  console.log(scrollWidth)
-  const clickHandle = () => {
-    containerRef.current.scrollBy({
-      left: scrollWidth,
-      behavior: "smooth",
-    });
-  };
-
-  return (
-    <div className={styles.container}>
-      <div className={styles.wrapper} onClick={clickHandle}>
-        {clickedYear && <ChevronDown height={120} width={120} fill="white" />}
-      </div>
-    </div>
-  );
-};
-
-export { RightCalendarArrow };
